Criar modal, ajuste de layout. Criar modal para tela de cadastrar cão, ajustas tela de cadastrar cão, criar toast para criação do cão.
