+++ src/app/cao/alterar-cao/alterar-cao.page.html
@@ -1,11 +1,9 @@
-<ion-header>
-  <h2 style="text-align: center;"><b>Alterar cão</b></h2>
-</ion-header>
-
 <ion-content>
   <div id="conteudo">
     <div id="inputs">
       
+      <h2 id="titulo"><b>Alterar cão</b></h2>
+
       <form #f="ngForm" (ngSubmit)="alterarCao(f)">
 
         <ion-item>
