+++ src/app/cao/alterar-cao/alterar-cao.page.scss
@@ -1,0 +1,14 @@
+#titulo{
+    text-align: center;
+    font-size: 120%;
+    font-family:arial;
+}
+
+#botao{
+    margin-bottom: 15px;
+
+    position: fixed;
+    left: 0;
+    bottom: 0px;
+    right: 0;
+}
