+++ src/app/cao/petsjacadastrados/petsjacadastrados.page.html
@@ -28,40 +28,6 @@
 
           </div>
 
-          <!-- <div id=dado>
-            <b>Nome: &nbsp;</b>
-            <b>Idade: </b>
-            <b>Peso: </b>
-            <b>Raça: </b>
-            <b>Porte: </b>
-          </div> -->
-          
-          <!-- <div id="informacao">
-            <div>
-              {{c.nome}}
-            </div>
-            <div *ngIf="c.idade > 1">
-              {{c.idade}} anos
-            </div>
-            <div *ngIf="c.idade == 1">
-              {{c.idade}} ano
-            </div>
-            <div>
-              {{c.peso.descricao}}
-            </div>
-            <div>
-              {{c.raca}}
-            </div>
-            <div *ngIf="c.porte == 1">
-              Pequeno
-            </div>
-            <div *ngIf="c.porte == 2">
-              Médio
-            </div>
-            <div *ngIf="c.porte == 3">
-              Grande
-            </div>
-          </div> -->
         </section>
   
         <section id="botoes">
