+++ src/app/menu/meuscaes/meuscaes.page.scss
@@ -2,21 +2,9 @@
 //    --ion-color-primary: rgb(90, 187, 0) !important;} //exemplo de como mudar a cor de algo
 //é só declarar a classe class na frente do item, e mexer aqui 
 
-.tituloToolBar{
-    font-family:arial;
-}
 .cadastreSeusCaesAqui{
     text-align: center;
     font-size: 120%;
-    font-family:arial;
-}
-
-
-.botaoCadastrarPet{
-    font-family:arial;
-}
-
-.verPetsJaCadastrados{
     font-family:arial;
 }
 
@@ -25,3 +13,17 @@
     width:  100%;
 }
 
+#title{
+    text-align: center;
+    font-size: 120%;
+    font-family:arial;
+}
+
+.botaoCadastrarPet{
+    margin-bottom: 15px;
+
+    position: fixed;
+    left: 0;
+    bottom: 0px;
+    right: 0;
+}
